nkids variable type check

+++ minos/modules/ageing.py
@@ -61,7 +61,8 @@ class Ageing(Base):
         logging.info(f"Aged population to year {event.time.year}")
 
         #if type(population['nkids'][0]) == float:
-        population['nkids'] = population['nkids'].astype(float)
+        #population['nkids'] = population['nkids'].astype(float)
+        #population['nkids'] = population['nkids'].astype(int)
         self.population_view.update(population[['age', 'time', 'child_ages', 'nkids']])
